+++ app/Http/Controllers/MainApi/PostController.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\MainApi;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use App\Http\Controllers\MainController;
 
 class PostController extends MainController
 {
